+++ src/main/java/snownee/researchtable/plugin/grandeconomy/CriterionMoneyGE.java
@@ -12,9 +12,9 @@ import snownee.researchtable.core.ICriterion;
 
 public class CriterionMoneyGE implements ICriterion
 {
-    private final long money;
+    private final double money;
 
-    public CriterionMoneyGE(long money)
+    public CriterionMoneyGE(double money)
     {
         this.money = money;
     }
@@ -32,11 +32,12 @@ public class CriterionMoneyGE implements ICriterion
     public String getFailingText(EntityPlayer player, NBTTagCompound data)
     {
         NBTHelper helper = NBTHelper.of(data);
-        return I18n.format(ResearchTable.MODID + ".gui.needMoney", money(money, helper), Util.color(0xFF0000) + I18n.format(ResearchTable.MODID + ".gui.youHave", money(helper.getLong("grandeconomy.money"), helper)));
+        return I18n.format(ResearchTable.MODID + ".gui.needMoney", money(money, helper), Util.color(0xFF0000) + I18n
+                .format(ResearchTable.MODID + ".gui.youHave", money(helper.getLong("grandeconomy.money"), helper)));
     }
 
     @SideOnly(Side.CLIENT)
-    public String money(long amount, NBTHelper helper)
+    public String money(double amount, NBTHelper helper)
     {
         String key = "grandeconomy." + (amount == 1 ? "singular" : "multiple");
         return I18n.format(ResearchTable.MODID + ".gui.moneyFormat", amount, helper.getString(key, ""));
